Compute the exact string length for IMAP commands
By computing the maximum length and then using `strlen()` we were being
efficient, but if there was a `\0` in the command (e.g. when there is an
attachment in a message and its data has that byte), `strlen()` skipped
the rest of the string and returned a smaller length.

Suggested `NEWS` line:
* `Avoid server freeze due to attached files`