feat(smtp): Support smtps
Implement support for SMTPS. Remove some unused initialization
functions and add one expecting NSURL. From this url we can extract
the scheme: smtp:// or smtps://, whereas the former is unencrypted SMTP
and the latter refers to smtp in SSL/TLS wrapper mode (usually port 465).

As with IMAPS and STARTTLS for sieve, openssl or GnuTLS need to be present
for this to work.

Fixes #31