feat(core): Support older GnuTLS versions
Add version checking to GnuTLS and use manual host name verification
in case GnuTLS does not support this natively.

Because older GnuTLS versions do not support using the system trust,
add a new configuration option to pass the system CA bundle. This option
can also search for the default locations. If none is given, but the
the GnuTLS version does not support loading the default system trust,
compilation will fail.