feat(core): NGActiveSSLSocket: Allow setting peer verification for STARTTLS
Add a new initializer for NGActiveSSLSocket, `initWithConnectedActiveSocket`
with the purpose to be initialized on top of a connected socket. Moving
the function of grabbing the socket fd and options from outside into the
SSL socket implementation itself.

Also by passing the socket, we dont have to pass the host name explicitly
anymore, as we can get it from the socket itself.

The verifyMode selector can enable the full TLS checks (default), disable
them for localhost addresses, or disable peer verification completely.

Refs #5078