fix(core): Dont leak fd when using STARTTLS
When STARTTLS is used, we re-use the connection and put a
new socket "on top" of the old one and reuse the file
descriptor of the previous socket.

This lead to two issues: First, now both sockets tried to close
the underlying file descriptor: Once from the non-starttls connection,
once by the starttls connection. This wasnt so harmful, as closing would
fail with EBADF. Fix this by using the flag `closeOnFree` for such
fd assignments, in this way the ownership is clearer.

The second issue is more severe: when passing the fd to the TLS socket,
the TLS socket could already have an fd assigned, but that would never
be freed. To fix the fd leak simply call `close` on the old fd
when `setFileDescriptor` is called.

Fixes: #5175